+++ cmd/hidden_lake/helpers/loader/internal/handler/handler_test.go
@@ -2,6 +2,7 @@ package handler
 
 import (
 	"net/http"
+	"time"
 
 	"github.com/number571/go-peer/cmd/hidden_lake/helpers/loader/internal/config"
 	"github.com/number571/go-peer/cmd/hidden_lake/helpers/loader/pkg/settings"
@@ -36,8 +37,9 @@ func testRunService(addr string) *http.Server {
 	mux.HandleFunc(settings.CHandleTransferPath, HandleTransferAPI(cfg, logger))
 
 	srv := &http.Server{
-		Addr:    addr,
-		Handler: mux,
+		Addr:        addr,
+		ReadTimeout: time.Second,
+		Handler:     http.TimeoutHandler(mux, time.Minute/2, "timeout"),
 	}
 
 	go func() {